Move DatasetStatusRead from `models` to `schemas`
