+++ fractal_server/app/api/v1/dataset.py
@@ -14,12 +14,12 @@ from ...db import AsyncSession
 from ...db import get_db
 from ...models import ApplyWorkflow
 from ...models import Dataset
-from ...models import DatasetStatusRead
 from ...models import JobStatusType
 from ...models import Resource
 from ...runner._common import HISTORY_FILENAME
 from ...schemas import DatasetCreate
 from ...schemas import DatasetRead
+from ...schemas import DatasetStatusRead
 from ...schemas import DatasetUpdate
 from ...schemas import ResourceCreate
 from ...schemas import ResourceRead
