+++ fractal_server/app/models/dataset.py
@@ -1,7 +1,6 @@
 from typing import Any
 from typing import Optional
 
-from pydantic import BaseModel
 from sqlalchemy import Column
 from sqlalchemy.types import JSON
 from sqlmodel import Field
@@ -10,26 +9,11 @@ from sqlmodel import SQLModel
 
 from ..schemas.dataset import _DatasetBase
 from ..schemas.dataset import _ResourceBase
-from ..schemas.workflow import WorkflowTaskStatusType
 
 
 class Resource(_ResourceBase, SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     dataset_id: int = Field(foreign_key="dataset.id")
-
-
-class DatasetStatusRead(BaseModel):
-    """
-    Response type for the
-    `/project/{project_id}/dataset/{dataset_id}/status/` endpoint
-    """
-
-    status: Optional[
-        dict[
-            int,
-            WorkflowTaskStatusType,
-        ]
-    ] = None
 
 
 class Dataset(_DatasetBase, SQLModel, table=True):
